Send br and arenas embeds instead of ranked and pubs`
